Basica 1.6.2 update

+++ src/client/olddays/behavior/EntityAIPanic.java
@@ -5,12 +5,12 @@ public class EntityAIPanic extends EntityAIBase
     public static boolean disablePanic = false;
 
     private EntityCreature theEntityCreature;
-    private float speed;
+    private double speed;
     private double randPosX;
     private double randPosY;
     private double randPosZ;
 
-    public EntityAIPanic(EntityCreature par1EntityCreature, float par2)
+    public EntityAIPanic(EntityCreature par1EntityCreature, double par2)
     {
         theEntityCreature = par1EntityCreature;
         speed = par2;
